Refactored panels code to multiple files.
- Added Time struct.
