+++ Humble2/src/Humble2/Core/Application.h
@@ -5,6 +5,7 @@
 
 #include "Base.h"
 #include "Input.h"
+#include "Time.h"
 #include "Context.h"
 #include "EventDispatcher.h"
