+++ Humble2/src/Humble2/Core/Time.cpp
@@ -1,0 +1,7 @@
+#include "Time.h"
+
+namespace HBL2
+{
+	float Time::DeltaTime = 0.0f;
+	float Time::FixedDeltaTime = 0.0f;
+}
