+++ Humble2/src/Humble2/Core/Time.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "Humble2API.h"
+
+namespace HBL2
+{
+	struct HBL2_API Time
+	{
+		static float DeltaTime;
+		static float FixedDeltaTime;
+	};
+}
