+++ HumbleEditor/src/Panels/ConsolePanel.cpp
@@ -1,0 +1,12 @@
+#include "Systems\EditorPanelSystem.h"
+
+namespace HBL2
+{
+	namespace Editor
+	{
+		void EditorPanelSystem::DrawConsolePanel(float ts)
+		{
+			ImGui::Text("Console");
+		}
+	}
+}
