+++ HumbleEditor/src/Panels/StatsPanel.cpp
@@ -1,0 +1,12 @@
+#include "Systems\EditorPanelSystem.h"
+
+namespace HBL2
+{
+	namespace Editor
+	{
+		void EditorPanelSystem::DrawStatsPanel(float ts)
+		{
+			ImGui::Text("Frame Time: %f", ts);
+		}
+	}
+}
