+++ HumbleEditor/src/Panels/ToolBarPanel.cpp
@@ -1,0 +1,216 @@
+#include "Systems\EditorPanelSystem.h"
+
+#include <Utilities\FileDialogs.h>
+
+namespace HBL2
+{
+	namespace Editor
+	{
+		void EditorPanelSystem::DrawToolBarPanel()
+		{
+			if (ImGui::BeginMainMenuBar())
+			{
+				if (ImGui::BeginMenu("File"))
+				{
+					if (ImGui::MenuItem("New Project"))
+					{
+						std::string filepath = HBL2::FileDialogs::SaveFile("Humble Project (*.hblproj)\0*.hblproj\0");
+
+						std::string projectName = std::filesystem::path(filepath).filename().stem().string();
+
+						AssetManager::Instance->DeregisterAssets();
+
+						HBL2::Project::Create(projectName)->Save(filepath);
+
+						auto assetHandle = AssetManager::Instance->CreateAsset({
+							.debugName = "Empty Scene",
+							.filePath = HBL2::Project::GetActive()->GetSpecification().StartingScene,
+							.type = AssetType::Scene,
+						});
+
+						HBL2::Project::OpenStartingScene();
+
+						m_ProjectChanged = true;
+						m_EditorScenePath = HBL2::Project::GetAssetFileSystemPath(HBL2::Project::GetActive()->GetSpecification().StartingScene);
+						m_CurrentDirectory = HBL2::Project::GetAssetDirectory();
+					}
+					if (ImGui::MenuItem("Open Project"))
+					{
+						std::string filepath = HBL2::FileDialogs::OpenFile("Humble Project (*.hblproj)\0*.hblproj\0");
+
+						AssetManager::Instance->DeregisterAssets();
+
+						if (HBL2::Project::Load(std::filesystem::path(filepath)) != nullptr)
+						{
+							HBL2::Project::OpenStartingScene();
+
+							m_ProjectChanged = true;
+							m_EditorScenePath = HBL2::Project::GetAssetFileSystemPath(HBL2::Project::GetActive()->GetSpecification().StartingScene);
+							m_CurrentDirectory = HBL2::Project::GetAssetDirectory();
+						}
+						else
+						{
+							HBL2_ERROR("Could not open specified project at path \"{0}\".", filepath);
+						}
+					}
+					if (ImGui::MenuItem("Save Scene"))
+					{
+						if (m_EditorScenePath.empty())
+						{
+							std::string filepath = HBL2::FileDialogs::SaveFile("Humble Scene (*.humble)\0*.humble\0");
+							auto relativePath = std::filesystem::relative(std::filesystem::path(filepath), HBL2::Project::GetAssetDirectory());
+
+							AssetManager::Instance->SaveAsset(std::hash<std::string>()(relativePath.string()));
+
+							m_EditorScenePath = filepath;
+						}
+						else
+						{
+							auto relativePath = std::filesystem::relative(m_EditorScenePath, HBL2::Project::GetAssetDirectory());
+
+							AssetManager::Instance->SaveAsset(std::hash<std::string>()(relativePath.string()));
+						}
+					}
+					if (ImGui::MenuItem("Save Scene As"))
+					{
+						std::string filepath = HBL2::FileDialogs::SaveFile("Humble Scene (*.humble)\0*.humble\0");
+						auto relativePath = std::filesystem::relative(std::filesystem::path(filepath), HBL2::Project::GetAssetDirectory());
+						const char* sceneName = relativePath.filename().stem().string().c_str();
+
+						auto assetHandle = AssetManager::Instance->CreateAsset({
+							.debugName = sceneName,
+							.filePath = relativePath,
+							.type = AssetType::Scene,
+						});
+
+						AssetManager::Instance->SaveAsset(assetHandle);
+						Asset* asset = AssetManager::Instance->GetAssetMetadata(assetHandle);
+
+						Handle<Scene> sceneHandle = Handle<Scene>::UnPack(asset->Indentifier);
+
+						Scene* playScene = ResourceManager::Instance->GetScene(sceneHandle);
+						Scene::Copy(m_ActiveScene, playScene);
+						AssetManager::Instance->SaveAsset(assetHandle);
+
+						SceneManager::Get().LoadScene(assetHandle);
+
+						m_EditorScenePath = filepath;
+					}
+					if (ImGui::MenuItem("New Scene"))
+					{
+						std::string filepath = HBL2::FileDialogs::SaveFile("Humble Project (*.humble)\0*.humble\0");
+						auto relativePath = std::filesystem::relative(std::filesystem::path(filepath), HBL2::Project::GetAssetDirectory());
+
+						auto assetHandle = AssetManager::Instance->CreateAsset({
+							.debugName = "New Scene",
+							.filePath = relativePath,
+							.type = AssetType::Scene,
+						});
+
+						AssetManager::Instance->SaveAsset(std::hash<std::string>()(relativePath.string()));
+
+						HBL2::SceneManager::Get().LoadScene(assetHandle);
+
+						m_EditorScenePath = filepath;
+					}
+					if (ImGui::MenuItem("Open Scene"))
+					{
+						std::string filepath = HBL2::FileDialogs::OpenFile("Humble Project (*.humble)\0*.humble\0");
+						auto relativePath = std::filesystem::relative(std::filesystem::path(filepath), HBL2::Project::GetAssetDirectory());
+						UUID sceneUUID = std::hash<std::string>()(relativePath.string());
+
+						HBL2::SceneManager::Get().LoadScene(AssetManager::Instance->GetHandleFromUUID(sceneUUID));
+
+						m_EditorScenePath = filepath;
+					}
+					if (ImGui::MenuItem("Build (Windows - Debug)"))
+					{
+						const std::string& projectName = HBL2::Project::GetActive()->GetSpecification().Name;
+
+						// Build.
+						system("\"C:\\Program Files\\Microsoft Visual Studio\\2022\\Community\\MSBuild\\Current\\Bin\\msbuild.exe\" ..\\HumbleGameEngine2.sln /t:HumbleApp /p:Configuration=Debug");
+
+						// Copy assets to build folder.
+						std::filesystem::copy("./" + projectName, "..\\bin\\Debug-x86_64\\HumbleApp\\" + projectName, std::filesystem::copy_options::recursive | std::filesystem::copy_options::overwrite_existing);
+
+						// Copy assets to build folder.
+						std::filesystem::copy("./assets", "..\\bin\\Debug-x86_64\\HumbleApp\\assets", std::filesystem::copy_options::recursive | std::filesystem::copy_options::overwrite_existing);
+					}
+					if (ImGui::MenuItem("Build (Windows - Release)"))
+					{
+						const std::string& projectName = HBL2::Project::GetActive()->GetSpecification().Name;
+
+						// Build.
+						system("\"C:\\Program Files\\Microsoft Visual Studio\\2022\\Community\\MSBuild\\Current\\Bin\\msbuild.exe\" ..\\HumbleGameEngine2.sln /t:HumbleApp /p:Configuration=Release");
+
+						// Copy assets to build folder.
+						std::filesystem::copy("./" + projectName, "..\\bin\\Release-x86_64\\HumbleApp\\" + projectName, std::filesystem::copy_options::recursive | std::filesystem::copy_options::overwrite_existing);
+
+						// Copy assets to build folder.
+						std::filesystem::copy("./assets", "..\\bin\\Release-x86_64\\HumbleApp\\assets", std::filesystem::copy_options::recursive | std::filesystem::copy_options::overwrite_existing);
+					}
+					if (ImGui::MenuItem("Build & Run (Windows - Debug)"))
+					{
+						const std::string& projectName = HBL2::Project::GetActive()->GetSpecification().Name;
+
+						// Build.
+						system("\"C:\\Program Files\\Microsoft Visual Studio\\2022\\Community\\MSBuild\\Current\\Bin\\msbuild.exe\" ..\\HumbleGameEngine2.sln /t:HumbleApp /p:Configuration=Debug");
+
+						// Copy assets to build folder.
+						std::filesystem::copy("./" + projectName, "..\\bin\\Debug-x86_64\\HumbleApp\\" + projectName, std::filesystem::copy_options::recursive | std::filesystem::copy_options::overwrite_existing);
+
+						// Copy assets to build folder.
+						std::filesystem::copy("./assets", "..\\bin\\Debug-x86_64\\HumbleApp\\assets", std::filesystem::copy_options::recursive | std::filesystem::copy_options::overwrite_existing);
+
+						// Run.
+						system("cd ..\\bin\\Debug-x86_64\\HumbleApp && HumbleApp.exe");
+					}
+					if (ImGui::MenuItem("Build & Run (Windows - Release)"))
+					{
+						const std::string& projectName = HBL2::Project::GetActive()->GetSpecification().Name;
+
+						// Build.
+						system("\"C:\\Program Files\\Microsoft Visual Studio\\2022\\Community\\MSBuild\\Current\\Bin\\msbuild.exe\" ..\\HumbleGameEngine2.sln /t:HumbleApp /p:Configuration=Release");
+
+						// Copy assets to build folder.
+						std::filesystem::copy("./" + projectName, "..\\bin\\Release-x86_64\\HumbleApp\\" + projectName, std::filesystem::copy_options::recursive | std::filesystem::copy_options::overwrite_existing);
+
+						// Copy assets to build folder.
+						std::filesystem::copy("./assets", "..\\bin\\Release-x86_64\\HumbleApp\\assets", std::filesystem::copy_options::recursive | std::filesystem::copy_options::overwrite_existing);
+
+						// Run.
+						system("cd ..\\bin\\Release-x86_64\\HumbleApp && HumbleApp.exe");
+					}
+					if (ImGui::MenuItem("Build (Web)"))
+					{
+						system("..\\Scripts\\emBuildAll.bat");
+					}
+					if (ImGui::MenuItem("Close"))
+					{
+						HBL2::Window::Instance->Close();
+					}
+					ImGui::EndMenu();
+				}
+
+				if (ImGui::BeginMenu("Edit"))
+				{
+					ImGui::EndMenu();
+				}
+
+				if (ImGui::BeginMenu("View"))
+				{
+					m_Context->GetRegistry()
+						.view<Component::EditorPanel>()
+						.each([&](Component::EditorPanel& panel)
+						{
+							ImGui::Checkbox(panel.Name.c_str(), &panel.Enabled);
+						});
+
+					ImGui::EndMenu();
+				}
+
+				ImGui::EndMainMenuBar();
+			}
+		}
+	}
+}
